# server/documentation/ONGAlkemy-Swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Api Rest for ONG",
    "version": "1.0.0",
    "title": "Proyect ONG Alkmey"
  },
  "host": "localhost:3000",
...
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/users/": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get all users",
        "description": "Get all users records",
        "operationId": "allUserList",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "description": "Token",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "User list found succesfully",
            "schema": {
              "$ref": "#/definitions/users"
            }
          },
          "401": {
            "description": "The token has expired"
          },
          "403": {
            "description": "Not have permission of admin"
          },
          "500": {
            "description": "Error with data base"
          }
        }
      }
    },
    "/users/auth/me": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Info personal",
        "description": "Get info personal",
        "operationId": "getPersonalInfo",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "description": "Token",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "true",
            "schema": {
              "$ref": "#/definitions/user"
            }
          },
          "401": {
            "description": "The token has expired"
          },
          "403": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/users/auth/register": {
      "post": {
        "tags": [
          "users"
        ],
        "summary": "Add a new user",
        "description": "Create new record user",
        "operationId": "userRegistration",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Data to complete register",
            "required": true,
            "schema": {
              "type": "object",
              "required": [
                "email",
                "password",
                "firstName",
                "lastName"
              ],
              "properties": {
                "password": {
                  "type": "string"
                },
                "email": {
                  "type": "string"
                },
                "firstName": {
                  "type": "string"
                },
                "lastName": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "true",
            "schema": {
              "$ref": "#/definitions/user"
            }
          },
          "400": {
            "description": "Bad request: The resource was not requested well"
          },
          "404": {
            "description": "El usuario ya a sido registrado"
          }
        }
      }
    },
    "/users/auth/login": {
      "post": {
        "tags": [
          "users"
        ],
        "summary": "Login user",
        "description": "Login user",
        "operationId": "authenticateLogin",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Data to complete register",
            "required": true,
            "schema": {
              "type": "object",
              "required": [
                "email",
                "password"
              ],
              "properties": {
                "password": {
                  "type": "string"
                },
                "email": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "true",
            "schema": {
              "$ref": "#/definitions/user"
            }
          },
          "404": {
            "description": "Not Found: Resource not found."
          }
        }
      }
    },
    "/users/{id}": {
      "delete": {
        "tags": [
          "users"
        ],
        "summary": "Delete user",
        "description": "Remove user",
        "operationId": "deleteUserById",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "Id to delete user",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "User Deleted"
          },
          "404": {
            "description": "Not Found: Resource not found."
          }
        }
      },
      "put": {
        "tags": [
          "users"
        ],
        "summary": "Update user",
        "description": "Update data user",
        "operationId": "userUpdate",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "Id to update user",
            "required": true,
            "type": "integer"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Data to complete update",
            "required": true,
            "schema": {
              "type": "object",
              "required": [
                "roleId",
                "firstName",
                "lastName"
              ],
              "properties": {
                "roleId": {
                  "type": "integer"
                },
                "firstName": {
                  "type": "string"
                },
                "lastName": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "true",
            "schema": {
              "$ref": "#/definitions/user"
            }
          },
          "401": {
            "description": "The user trying to update does not exist"
          },
          "403": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  },
  "definitions": {
    "users": {
      "type": "array",
      "items": {
        "$ref": "#/definitions/user"
      }
    },
...
    "user": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "image": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "roleId": {
          "type": "integer"
        },
        "deletedAt": {
          "type": "string"
        }
      },
      "xml": {
        "name": "user"
      },
      "required": [
        "id",
        "email",
        "password",
        "firstName",
        "roleId"
      ]
    },
...
  }
}